requirements.txt
streamlit>=1.28.0
pandas>=1.3.0
numpy>=1.21.0
plotly>=5.0.0
bcrypt>=4.0.1
Pillow>=9.0.0
scikit-learn>=1.1.0
nltk>=3.7
spacy>=3.4.0
wordcloud>=1.8.0
seaborn>=0.11.0
networkx>=2.8.0
umap-learn>=0.5.0
hdbscan>=0.8.0
sentence-transformers>=2.2.0
openai>=0.27.0
langchain>=0.0.200
chromadb>=0.3.0
transformers>=4.21.0
torch>=1.12.0
tensorflow>=2.9.0
scipy>=1.7.0
textblob>=0.17.1
vaderSentiment>=3.3.2
googletrans>=3.1.0a0
streamlit-chat>=0.1.1
streamlit-authenticator>=0.2.1
streamlit-option-menu>=0.3.6
streamlit-extras>=0.2.7
email-validator>=1.3.1
streamlit-aggrid>=0.3.0
streamlit-plotly-events>=0.0.6
streamlit-lottie>=0.0.3
